requirements.txt
click==8.1.6
loguru==0.7.0
google-cloud-compute==1.14.0
google-cloud-resource-manager==1.10.3
google-cloud-monitoring==2.15.1
google-cloud-redis==2.13.1
google-cloud-container==2.29.0
google-api-python-client==2.97.0
google-cloud-recommender==2.11.2
google-cloud-essential-contacts==1.5.3
google-cloud-storage==2.10.0